Added statistic support: count number of R0-RN, RE and BF applications.
[r24595]

+++ bucket.c
@@ -52,7 +52,7 @@ pbqp_edge *edge_bucket_pop(pbqp_edge_bucket *bucket)
 
 void node_bucket_shrink(pbqp_node_bucket *bucket, unsigned len)
 {
-	ARR_SHRINKLEN(*bucket, len);
+	ARR_SHRINKLEN(*bucket, (int)len);
 }
 
 int node_bucket_contains(pbqp_node_bucket bucket, pbqp_node *node)
